Updaed code

+++ assignment2/Routes/post.js
@@ -1,5 +1,6 @@
 const express = require("express");
 const Post = require("../Models/post");
+const User = require("../Models/user");
 
 const app = express();
 
@@ -7,7 +8,8 @@ app.use(express.json());
 
 app.get("/", async (req, res) => {
   try {
-    const posts = await Post.find();
+    const currentUser = await User.findOne({ id: req.user });
+    const posts = await Post.find({ name: currentUser.name });
     res.status(200).json({
       status: "Success",
       posts,
@@ -22,7 +24,13 @@ app.get("/", async (req, res) => {
 
 app.post("/", async (req, res) => {
   try {
-    const posts = await Post.create(req.body);
+    const currentUser = await User.findOne({ id: req.user });
+    const posts = await Post.create({
+      title: req.body.title,
+      body: req.body.body,
+      image: req.body.image,
+      user: currentUser.name,
+    });
     res.status(200).json({
       status: "Success",
       massage: "Post added successfully",
@@ -37,13 +45,26 @@ app.post("/", async (req, res) => {
 });
 
 app.put("/:postId", async (req, res) => {
-    try {
-        const posts = await Post.updateOne({ _id: req.params.postId }, req.body);
-        res.status(200).json({
-            status: "Success",
-            massage: "Post successfully updated",
-            posts
-        })
+  try {
+    const currentUser = await User.findOne({ id: req.user });
+    const posts = await Post.updateOne({ _id: req.params.postId,user:currentUser.name }, {
+      user:currentUser.name,
+      title: req.body.title,
+      image:req.body.image,
+      body:req.body.body
+    });
+    if (posts) {
+      res.status(200).json({
+        status: "Success",
+        massage: "Post successfully updated",
+        posts,
+      });
+    } else {
+      res.status(400).json({
+        status: "Failed",
+        massgae:"Cannot update other users post"
+      })
+    }
   } catch (e) {
     res.status(500).json({
       status: "Failed",
@@ -53,12 +74,20 @@ app.put("/:postId", async (req, res) => {
 });
 
 app.delete("/:postId", async (req, res) => {
-    try {
-        await Post.deleteOne({ _id: req.params.postId });
-        res.status(200).json({
-            status: "Success",
-            massage:"Post deleted successfully"
+  try {
+    const currentUser = await User.findOne({ id: req.user });
+    const posts = await Post.deleteOne({ _id: req.params.postId, user: currentUser.name });
+    if (posts) {
+      res.status(200).json({
+        status: "Success",
+        massage: "Post deleted successfully",
+      });   
+    } else {
+      res.status(400).json({
+        status: "Failed",
+        massage:"Cannot delete post of other users"
       })
+    }
   } catch (e) {
     res.status(500).json({
       status: "Failed",
